vae/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vae-torch"
version = "0.1.0"
description = "PyTorch reproduction of a modern convolutional VAE from scratch"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch",
  "torchvision",
  "datasets",
  "pyyaml",
  "pillow",
  "tqdm",
  "numpy",
  "pytest",
]

[project.scripts]
vae-prepare = "vae_torch.data:main"
vae-train = "vae_torch.train:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
